third_party/WebKit/LayoutTests/inspector/debugger/async-callstack-expected.txt

Tests asynchronous call stacks in debugger.

Debugger was enabled.
Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) animFrame1 (async-callstack.html:30)
    [Async Call]
        0) timeout1 (async-callstack.html:19)
        1) innerTestFunction (async-callstack.html:11)
    [Async Call]
        0) testFunction (async-callstack.html:13)

Call stack:
    0) animFrame2 (async-callstack.html:47)
    [Async Call]
        0) animFrame1 (async-callstack.html:32)
    [Async Call]
        0) timeout1 (async-callstack.html:19)
        1) innerTestFunction (async-callstack.html:11)
    [Async Call]
        0) testFunction (async-callstack.html:13)

Call stack:
    0) interval1 (async-callstack.html:37)
    1) innerInterval1 (async-callstack.html:24)
    [Async Call]
        0) timeout1 (async-callstack.html:20)
        1) innerTestFunction (async-callstack.html:11)
    [Async Call]
        0) testFunction (async-callstack.html:13)

Call stack:
    0) timeout1 (async-callstack.html:18)
    1) innerTestFunction (async-callstack.html:11)
    [Async Call]
        0) testFunction (async-callstack.html:13)

Call stack:
    0) timeout2 (async-callstack.html:42)
    [Async Call]
        0) animFrame1 (async-callstack.html:31)
    [Async Call]
        0) timeout1 (async-callstack.html:19)
        1) innerTestFunction (async-callstack.html:11)
    [Async Call]
        0) testFunction (async-callstack.html:13)

Call stack:
    0) timeout3 (async-callstack.html:61)
    1) longTail0 (async-callstack.html:50)
    [Async Call]
        0) longTail1 (:1)
    [Async Call]
        0) longTail2 (:2)
    [Async Call]
        0) longTail3 (:3)
    [Async Call]
        0) longTail4 (:4)

Debugger was disabled.
